move validations to separate file
