+++ src/PapersBee/papers/__init__.py
@@ -1,4 +1,5 @@
 from . import config
 from .papers_finder import PapersFinder
+from .validate_inputs import validate_configuration, validate_llm_args, validate_ncbi_api_key, validate_posting_args
 
-__all__ = ['PapersFinder', 'config']
+__all__ = ['PapersFinder', 'config', 'validate_configuration', 'validate_llm_args', 'validate_ncbi_api_key', 'validate_posting_args']
